Clean up query and scorer
